+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 - (breaking) Don't lower-case the `type` property on labels. `type` no longer has any special significance other than in the `label` embedded step. Its use should be replaced with `json_schema`.
+- Recognize short aliases for the json_schema label property. E.g., a label with `json_schema: boolean` is valid. The `boolean` alias corresponds to a JSON Schema that only allows boolean values.
+
 ## v0.13.0 (2023-01-24)
 
 - Add git revision to "sr version" output